backend/requirements.txt
flask
tensorflow==2.10.0
pillow
numpy